Reduce duplicate lint traversal work
