+++ crates/shuck-linter/src/checker.rs
@@ -530,9 +530,6 @@ impl<'a> Checker<'a> {
         if self.is_rule_enabled(Rule::CommentedContinuationLine) {
             rules::correctness::commented_continuation_line::commented_continuation_line(self);
         }
-        if self.is_rule_enabled(Rule::TrailingDirective) {
-            rules::style::trailing_directive::trailing_directive(self);
-        }
         if self.is_rule_enabled(Rule::UnicodeSingleQuoteInSingleQuotes) {
             rules::correctness::unicode_single_quote_in_single_quotes::unicode_single_quote_in_single_quotes(self);
         }
